+++ gradle-idea-configuration/src/main/groovy/com/palantir/gradle/ideaconfiguration/IdeaConfigurationExtension.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.gradle.gradleideaconfiguration;
+package com.palantir.gradle.ideaconfiguration;
 
 import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.Project;
