+++ gradle-idea-configuration/src/main/groovy/com/palantir/gradle/ideaconfiguration/PluginDependency.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.gradle.gradleideaconfiguration;
+package com.palantir.gradle.ideaconfiguration;
 
 import java.io.Serializable;
 import java.util.Objects;
