+++ gradle-idea-configuration/src/main/groovy/com/palantir/gradle/ideaconfiguration/UpdateExternalDependenciesXml.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.gradle.gradleideaconfiguration;
+package com.palantir.gradle.ideaconfiguration;
 
 import java.util.Set;
 import org.gradle.api.DefaultTask;
