+++ gradle-idea-configuration/src/main/groovy/com/palantir/gradle/ideaconfiguration/XmlUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.gradle.gradleideaconfiguration;
+package com.palantir.gradle.ideaconfiguration;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Files;
